OneDrive/Desktop/confession-backend/requirements.txt
Flask
Gunicorn
pillow
flask-cors
python-dotenv
uuid
instabot